completed environment variable setup and project bundling.
