+++ package.json
@@ -4,7 +4,8 @@
   "description": "Natawnee's first node.js javascript project.",
   "main": "natawneeProject.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --config webpack.config.js"
   },
   "repository": {
     "type": "git",
@@ -16,5 +17,14 @@
   "bugs": {
     "url": "https://github.com/tawneejoneskell/CatProject/issues"
   },
-  "homepage": "https://github.com/tawneejoneskell/CatProject#readme"
+  "homepage": "https://github.com/tawneejoneskell/CatProject#readme",
+  "devDependencies": {
+    "@babel/core": "^7.18.5",
+    "@babel/preset-env": "^7.18.2",
+    "babel-loader": "^8.2.5",
+    "dotenv": "^16.0.1",
+    "dotenv-webpack": "^7.1.0",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0"
+  }
 }
