+++ src/natawneeProject.html
@@ -3,7 +3,7 @@
 
 <head>
     <meta charset="ISO-8859-1">
-    <script src="./natawneeProject.js" ></script>
+    <script src="../dist/bundle.js" ></script>
     <link rel="stylesheet" href="./natawneeProject.css">
     <title> Choose your Cat!</title>
 </head>
